save2

+++ Assets/Script/Settings.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu]
 public class Settings : ScriptableObject
 {
-    public int HeroHealth = 100;
+    public float HeroHealth = 100;
     
 }
